Getting there... figured out exciting races between sub updates and awaiting txn result so refined semantics. Ref T64
--HG--
branch : T64

+++ client/cache.go
@@ -85,7 +85,7 @@ func (c *Cache) copyReferences(dstSeg *capn.Segment, refsAcc *[]*msgs.VarIdPos, 
 type Cached struct {
 	*Cache
 	VerClock
-	counter   uint32
+	counter   uint64
 	caps      common.Capability
 	refs      []msgs.VarIdPos
 	positions *common.Positions
